made SpectralDefinition subclass part of PlanetarySystem rather than Planet
